Code to test if the components are working
Test the functionality of the components (3 motors, GPS, temperature
sensor, pressure sensor) through serial port. To activate the motors,
send through serial:
"MOTmDXXX,", where: m= A, B, or C for each motor
D = L or R for direction
XXX =  a number between 0 and 255 for speed
To request temperature or pressure readings type "RQT," or "RQP," in
serial monitor. GPS reading can be received by using command "RQL,".
ALL commands must and with a "," for the instruction to work.