updated test code
Updated test code to include the analog multiplexer.
A0 and A1 are the select pins of the multiplexer and A6, A7 are the
analog input pins for the sensors. Writing LOW on both A0 and A1 the
first channel of the multiplexer is selected, where panstamps receives
input from the LM35 and the pressure sensors